requirements.txt
# Core dependencies
psycopg2-binary==2.9.7
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0

# Vector models and embeddings
sentence-transformers==2.2.2
scikit-learn==1.3.2
torch==2.1.1
transformers==4.36.2

# Web application
streamlit==1.29.0
plotly==5.17.0

# Data processing
openpyxl==3.1.2
xlrd==2.0.1

# Utilities
tqdm==4.66.1
requests==2.31.0

# Development and testing
pytest==7.4.3
black==23.11.0
flake8==6.1.0 